api/auth_views.py
"""
Authentication views for JWT tokens.
"""
from rest_framework import status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import AllowAny
from rest_framework.response import Response
from rest_framework_simplejwt.tokens import RefreshToken
from django.contrib.auth import authenticate
from drf_spectacular.utils import extend_schema


@extend_schema(
    summary="Obtain JWT token",
    description="Get access and refresh tokens for API authentication.",
    request={
        'application/json': {
            'type': 'object',
            'properties': {
                'username': {'type': 'string'},
                'password': {'type': 'string'},
            },
            'required': ['username', 'password'],
        },
    },
    responses={200: {'description': 'Tokens'}},
)
@api_view(['POST'])
@permission_classes([AllowAny])
def obtain_token(request):
    """Obtain JWT access and refresh tokens."""
    username = request.data.get('username')
    password = request.data.get('password')

    if not username or not password:
        return Response(
            {'error': 'Username and password are required'},
            status=status.HTTP_400_BAD_REQUEST,
        )

    user = authenticate(username=username, password=password)
    if not user:
        return Response(
            {'error': 'Invalid credentials'},
            status=status.HTTP_401_UNAUTHORIZED,
        )

    refresh = RefreshToken.for_user(user)
    return Response({
        'access': str(refresh.access_token),
        'refresh': str(refresh),
    })


@extend_schema(
    summary="Refresh JWT token",
    description="Refresh access token using refresh token.",
    request={
        'application/json': {
            'type': 'object',
            'properties': {
                'refresh': {'type': 'string'},
            },
            'required': ['refresh'],
        },
    },
    responses={200: {'description': 'New access token'}},
)
@api_view(['POST'])
@permission_classes([AllowAny])
def refresh_token(request):
    """Refresh JWT access token."""
    refresh_token = request.data.get('refresh')
    if not refresh_token:
        return Response(
            {'error': 'Refresh token is required'},
            status=status.HTTP_400_BAD_REQUEST,
        )

    try:
        refresh = RefreshToken(refresh_token)
        return Response({
            'access': str(refresh.access_token),
        })
    except Exception as e:
        return Response(
            {'error': 'Invalid refresh token'},
            status=status.HTTP_401_UNAUTHORIZED,
        )
