requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
clickhouse-connect==0.6.23
django-redis==5.4.0

# Async & Messaging
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Utilities
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2

# API Documentation
drf-spectacular==0.26.5

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
